Fix: Fix copy.js compare error
